UI System
I have begun working on a UI system that combines SFML and TGUI features
to create buttons specific for this project. Some of the previous
systems have been reworked, for example the ResourceManager and how I go
about creating and storing Textures and Sprites. I currently have the
TitleScreen implemented with the buttons working with mouse, however I
need to begin tracking which button is selected for compatibility with
keyboard / controller.