Game Loop
Implemented a proper game loop that calls the update() function on a
timer rather than calling it every frame. Implemented an FPS counter
that can be toggled with F11 (Got confused for a while as F12 is a VS
shortcut for triggering breakpoints). Also began implementing GUI
navigation for keyboard.