Title Screen Sub-States
Began implementing sub-states, these were originally going to be
seperate states from the title-screen, however I have realised that this
would prevent background animations and sounds from carrying over to
each of the different menus. My solution for this was the create
sub-states of the title screen so even when the user is within the
options menu, the currently active state is still the TitleScreenState.
The TitleScreenState now handles changing to new sub-states and moving
back to the main title screen. The sub-states can also have all the same
functionality as the TitleScreenState so sub-states act exactly like
states except they can be active at the same time as a state.