Key Binds View
Began learning how to use the sf::View class. The sf::View class is
necessary to create the small key binds menu that will display all of
the key binds to the player, obviously this list will be longer than the
screen so we make use of sf::View to make scrolling the text and buttons
fairly easy. As of now I am displaying all of the action names listed in
the XML files and these are scrollable and working fine however I will
need to limit the amount of scrolling possible. Next I need to create
the buttons that will display the currently bound key followed by the
rebind / unbind functionality.