Key Binds Buttons
The buttons have been implemented into the key bind menu. Now alongside
each 'action' name is a button that displays the currently bound key.
These buttons can be hovered over and clicked on with the mouse where
the 'clicked' on button will remain clicked on until the user clicks
another button. The buttons are also rendered on the previously
mentioned sf::View so scrolling is working on the buttons as well. Now
that the buttons are implemented I plan on adding the 'Rebind' 'Unbind'
buttons that will allow the user to alter the currently bound keys.